requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
build==1.2.2.post1
cachetools==5.5.2
camelot-py==0.11.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dashscope==1.22.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.9
et_xmlfile==2.0.0
etcd3==0.12.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
ghostscript==0.7
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kubernetes==32.0.1
llama-index==0.12.25
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.25
llama-index-embeddings-huggingface==0.5.2
llama-index-embeddings-ollama==0.6.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-dashscope==0.3.1
llama-index-llms-ollama==0.5.3
llama-index-llms-openai==0.3.26
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-readers-notion==0.3.0
llama-index-readers-pdf-table==0.3.0
llama-index-readers-smart-pdf-loader==0.3.0
llama-index-tools-yahoo-finance==0.3.0
llama-index-vector-stores-chroma==0.4.1
llama-index-vector-stores-milvus==0.5.0
llama-parse==0.6.4.post1
-e git+ssh://git@github.com/yocover/llamaindex-learning.git@c609da4fa4a004e87527d33e234dc22407932423#egg=llamaindex_learning
llmsherpa==0.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.12
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.2.0
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.21.0
openai==1.68.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20240706
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
posthog==3.21.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymilvus==2.5.6
pyparsing==3.2.1
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.18.3
yfinance==0.2.55
zipp==3.21.0
